rocket_league_ai_backend/requirements.txt
azure
azure-storage
django
djangorestframework
django-rest-swagger